chore: update imports after messaging refactor
