+++ tests/test_alert_webhook.py
@@ -8,7 +8,8 @@ import logging
 import pytest
 
 from alpha_factory_v1.core import orchestrator
-from alpha_factory_v1.core.utils import alerts, messaging, config
+from alpha_factory_v1.core.utils import alerts, config
+from alpha_factory_v1.common.utils import messaging
 from alpha_factory_v1.core.agents.base_agent import BaseAgent
